+++ dev-site-config/site.config.js
@@ -2,9 +2,8 @@ const navConfig = require('./navigation.config');
 
 module.exports = {
   navConfig,
-  filterSideMenu: true,
+  includeTestEvidence: false,
   sideEffectImports: [
     './initalizeXFC.js',
   ],
-  includeTestEvidence: false,
 };