+++ packages/terra-abstract-modal/src/terra-dev-site/doc/abstract-modal/AbstractModal.doc.jsx
@@ -14,6 +14,8 @@ import AbstractModalCloseOnOutsideClick from '../example/AbstractModalCloseOnOut
 import AbstractModalCloseOnOutsideClickSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/AbstractModalCloseOnOutsideClick';
 import AbstractModalIsFullscreen from '../example/AbstractModalIsFullscreen';
 import AbstractModalIsFullscreenSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/AbstractModalIsFullscreen';
+import AbstractModalAlertDialog from '../example/AbstractModalAlertDialog';
+import AbstractModalAlertDialogSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/AbstractModalAlertDialog';
 
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
@@ -37,6 +39,11 @@ const DocPage = () => (
         title: 'Fullscreen Modal',
         example: <AbstractModalIsFullscreen />,
         source: AbstractModalIsFullscreenSrc,
+      },
+      {
+        title: 'Alert Dialog Modal',
+        example: <AbstractModalAlertDialog />,
+        source: AbstractModalAlertDialogSrc,
       },
     ]}
     propsTables={[