+++ packages/terra-application-navigation/src/terra-dev-site/test/application-navigation/HeroDrawer.test.jsx
@@ -1,0 +1,16 @@
+/* eslint-disable */
+import React from 'react';
+import ApplicationBase from 'terra-application/lib/application-base';
+import ApplicationNavigation from '../../../ApplicationNavigation';
+import Placeholder from 'terra-doc-template/lib/Placeholder';
+
+const hero = <Placeholder title="Test Hero" variant="light" />;
+
+// TODO: remove terra-application after it is incorporated into dev-site
+const TitleTest = () => (
+  <ApplicationBase locale="en">
+    <ApplicationNavigation hero={hero} />
+  </ApplicationBase>
+);
+
+export default TitleTest;