+++ packages/terra-application-navigation/tests/jest/extensions/ExtensionCount.test.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import ExtensionCount from '../../../src/extensions/_ExtensionCount';
+
+describe('ExtensionCount', () => {
+  it('should render default element', () => {
+    const shallowComponent = shallow(
+      <ExtensionCount />,
+    );
+
+    expect(shallowComponent).toMatchSnapshot();
+  });
+
+  it('should render prop data', () => {
+    const shallowComponent = shallow(
+      <ExtensionCount
+        value={1}
+        isRollup
+      />,
+    );
+
+    expect(shallowComponent).toMatchSnapshot();
+  });
+});