+++ packages/terra-application-navigation/tests/jest/extensions/Extensions.test.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { mockIntl } from 'terra-enzyme-intl';
+import Extensions from '../../../src/extensions/_Extensions';
+
+describe('Extensions', () => {
+  it('should render default element', () => {
+    const shallowComponent = shallow(
+      <Extensions.WrappedComponent intl={mockIntl} />,
+    );
+
+    expect(shallowComponent).toMatchSnapshot();
+  });
+
+  it('should render prop data', () => {
+    const shallowComponent = shallow(
+      <Extensions.WrappedComponent
+        intl={mockIntl}
+        activeBreakpoint="small"
+        extensionItems={[{ text: 'test-text', icon: <span>my icon</span>, key: 'my-test-key' }]}
+        notifications={{ myComponent: 3 }}
+        onSelect={jest.fn()}
+      />,
+    );
+
+    expect(shallowComponent).toMatchSnapshot();
+  });
+});