+++ packages/terra-form-validation/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
 
 1.5.0 - (July 30, 2019)
 ------------------