+++ packages/terra-form-validation/CHANGELOG.md
@@ -4,6 +4,26 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
+* updated package.json test scripts
+* Added `noValidate` attribute to all `<form />` elements in the examples
+* Added documentation on recommend usage of `noValidate` prop
+
+1.5.0 - (July 30, 2019)
+------------------
+### Removed
+* Removed DEPENDENCIES.md file
+
+### Changed
+* Update main key in package.json to point to `lib/FormValidationUtil.js`
+
+1.4.1 - (July 23, 2019)
+------------------
+### Changed
+* Updated DEPENDENCIES.md doc
+
+1.4.0 - (July 23, 2019)
+------------------
+### Changed
 * Updated terra-date-picker devDependency to v4.7.0
 
 ### Removed